# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A Web3 backend for Sukuk (Islamic bonds) on Base Testnet, providing APIs for dApps and frontends while processing blockchain events from an indexer.",
        "title": "Sukuk POC Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.sukuk-poc.com/support",
            "email": "support@sukuk-poc.com"
        },
        "license": {
            "name": "MIT",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/admin/sukuks": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new Sukuk series with off-chain data (admin only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sukuk Management"
                ],
                "summary": "Create new Sukuk series",
                "parameters": [
                    {
                        "description": "Sukuk series data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.CreateSukukSeriesRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Sukuk series created",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/sukuks/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update existing Sukuk series off-chain data (admin only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sukuk Management"
                ],
                "summary": "Update Sukuk series",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Sukuk Series ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.UpdateSukukSeriesRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sukuk series updated",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Sukuk series not found",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/sukuks/{id}/upload-prospectus": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Upload PDF prospectus file for a Sukuk series (admin only)",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sukuk Management"
                ],
                "summary": "Upload Sukuk prospectus",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Sukuk Series ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "PDF prospectus file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Prospectus uploaded",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.FileUploadResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Sukuk series not found",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_handlers.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string",
                    "example": "Validation failed"
                },
                "message": {
                    "type": "string",
                    "example": "Operation completed successfully"
                }
            }
        },
...
        "internal_handlers.CreateSukukSeriesRequest": {
            "type": "object",
            "required": [
                "company_id",
                "maturity_date",
                "min_investment",
                "name",
                "symbol",
                "total_supply",
                "yield_rate"
            ],
            "properties": {
                "company_id": {
                    "type": "integer",
                    "example": 1
                },
                "description": {
                    "type": "string",
                    "example": "Sustainable infrastructure financing sukuk"
                },
                "is_redeemable": {
                    "type": "boolean",
                    "example": true
                },
                "maturity_date": {
                    "type": "string",
                    "example": "2027-12-31T00:00:00Z"
                },
                "max_investment": {
                    "type": "string",
                    "example": "100000000000000000000"
                },
                "min_investment": {
                    "type": "string",
                    "example": "1000000000000000000"
                },
                "name": {
                    "type": "string",
                    "example": "Green Sukuk Series A"
                },
                "payment_frequency": {
                    "type": "integer",
                    "example": 4
                },
                "symbol": {
                    "type": "string",
                    "example": "GSA"
                },
                "total_supply": {
                    "type": "string",
                    "example": "1000000000000000000000000"
                },
                "yield_rate": {
                    "type": "number",
                    "example": 0.085
                }
            }
        },
...
        "internal_handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Invalid request data"
                }
            }
        },
...
        "internal_handlers.FileUploadResponse": {
            "type": "object",
            "properties": {
                "filename": {
                    "type": "string",
                    "example": "company_logo_1.png"
                },
                "message": {
                    "type": "string",
                    "example": "File uploaded successfully"
                },
                "url": {
                    "type": "string",
                    "example": "/uploads/logos/company_logo_1.png"
                }
            }
        },
...
        "internal_handlers.UpdateSukukSeriesRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Updated description"
                },
                "name": {
                    "type": "string",
                    "example": "Updated Green Sukuk Series A"
                },
                "outstanding_supply": {
                    "type": "string",
                    "example": "1000000000000000000000000"
                },
                "status": {
                    "type": "string",
                    "example": "active"
                },
                "token_address": {
                    "type": "string",
                    "example": "0xabcdefabcdefabcdefabcdefabcdefabcdefabcd"
                },
                "total_supply": {
                    "type": "string",
                    "example": "2000000000000000000000000"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API key for accessing protected admin endpoints",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        }
    }
}